Check if an icon set is installed.

This function will return TRUE if the icons for an icon set are installed. If they aren't, they can be installed using the appropriate download_*() function.

icon_installed(x)

Arguments

x

An icon set (such as fontawesome).